+++ packages/components/molecules/f-alert/test/accessibility/axe-accessibility.spec.js
@@ -2,7 +2,7 @@ const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 const Alert = require('../../test-utils/component-objects/f-alert.component');
 
-const alert = new Alert('molecule', 'alert-component');
+const alert = new Alert();
 
 describe('Accessibility tests', () => {
     beforeEach(() => {