+++ packages/components/molecules/f-mega-modal/test-utils/component-objects/f-mega-modal.component.js
@@ -1,6 +1,10 @@
 const Page = require('@justeat/f-wdio-utils/src/page.object');
 
 module.exports = class MegaModal extends Page {
+    constructor() {
+        super('molecule', 'mega-modal-component');
+    }
+
     get component () { return $('[data-test-id="mega-modal-component"]'); }
     get megaModalTitle () { return $('[data-test-id="mega-modal-title"]'); }
     get megaModalContent () { return $('[data-test-id="mega-modal-content"]'); }