+++ packages/services/f-analytics/src/store/analytics.module.js
@@ -20,7 +20,8 @@ export default {
             jeUserPercentage: undefined,
             currency: undefined,
             version: undefined,
-            instancePosition: undefined
+            instancePosition: undefined,
+            isPilot: undefined
         },
         userData: {
             'a-UserId': undefined,
@@ -34,7 +35,6 @@ export default {
             group: undefined,
             name: undefined,
             httpStatusCode: 200,
-            isCached: false,
             conversationId: undefined,
             requestId: undefined,
             orientation: undefined,