+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,105 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.180.2
+------------------------------
+*September 8, 2021*
+
+### Added
+- Get checkout demo files for AU/NZ
+
+### Changed
+- Moved demo files inside the stories directory
+
+
+v0.180.1
+------------------------------
+*September 7, 2021*
+
+### Changed
+- Address labels for AU/NZ
+
+
+v0.180.0
+------------------------------
+*September 6, 2021*
+
+### Added
+- `administrativeArea` input for AU
+
+
+v0.179.0
+------------------------------
+*September 3, 2021*
+
+### Changed
+- `AgeVerification` page to show correct days per month.
+
+
+v0.178.0
+------------------------------
+*September 3, 2021*
+
+### Changed
+- Redirect to search or home if restaurant is not taking orders.
+- Redirect to menu if service is not available.
+
+
+v0.177.0
+------------------------------
+*September 1, 2021*
+
+### Changed
+- Gelocation warning to error for more details in Sentry
+
+
+v0.176.0
+------------------------------
+*August 25, 2021*
+
+### Changed
+- Map checkout update request now trims whitespace from postcode
+
+
+v0.175.0
+------------------------------
+*August 25, 2021*
+
+### Removed
+- Age verification div wrapper from `AgeVerification.vue` component.
+
+### Changed
+- `AgeVerificationError` ref location from wrapper to `errorMessage` component.
+- `handleAgeVerifcation` method to look for correct component when error appears.
+
+### Fixed
+- Typo in `AgeVerification.vue` component.
+
+
+v0.174.0
+------------------------------
+*August 25, 2021*
+
+### Added
+- Add copy for AU/NZ locales
+- Guest dine in visual test.
+
+
+v0.173.4
+------------------------------
+*August 25, 2021*
+
+### Changed
+- Increased timeout to 60s.
+
+
+v0.173.3
+------------------------------
+*August 24, 2021*
+
+### Added
+- Logging items on invalid checkout.
+
 
 v0.173.2
 ------------------------------
@@ -29,13 +128,6 @@ v0.173.0
 
 ### Changed
 - `checkoutMocks.js` & `checkout.stories.js` refactored substantially.
-
-(To be added to next release)
-------------------------------
-*August 12, 2021*
-
-### Added
-- Guest dine in visual test.
 
 
 v0.172.0