+++ packages/components/organisms/f-cookie-banner/CHANGELOG.md
@@ -3,6 +3,24 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.25.0
+------------------------------
+*September 9, 2021*
+
+### Added
+- Tests to cover version changes in v0.24.0.
+
+### Changed
+- `dk-DK` to `da-DK`. 
+
+
+v0.24.0
+------------------------------
+*August 27, 2021*
+
+### Fixed
+- Moved `isBodyHeightLessThanWindowHeight` to computed so it can re-calculate cookie position.
+
 
 v0.23.0
 ------------------------------