+++ packages/components/organisms/f-loyalty/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.4.0
+------------------------------
+*August 24, 2021*
+
+### Added
+- Tabs to show how it works
+- How it works page
+- Unit tests for how it works page
+
 
 v0.3.0
 ------------------------------